Jewelify_server/requirements.txt
# pip==25.0.1
fastapi==0.115.10
uvicorn==0.34.0
tensorflow==2.18.0
numpy==1.26.4
pillow==11.1.0
python-multipart==0.0.20
nest_asyncio==1.6.0
python-dotenv==1.0.1
pydantic==2.10.6
scikit-learn==1.5.2
pymongo==4.11.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0  # Updated to include cryptography backend
pyyaml==6.0.1
requests==2.32.3
bcrypt==4.2.0
xgboost==2.1.1
twilio==9.5.0
opencv-python==4.11.0.86
aiohttp==3.10.5
motor==3.7.0